+++ src/2d_engine/backend_api.hpp
@@ -1,0 +1,11 @@
+
+#pragma once
+
+namespace cv {
+
+    struct s32x2 {
+        s32 x;
+        s32 y;
+    };
+
+} // cv